Moved logic common to timeline to a new abstract class called TimelineFragment
Problem encountered: After composing a tweet i struggled with updating the correct
fragment because now the fragments are created by the fragmentAdapter, so i don't
have direct access to the fragment instances anymore.

Solution: Get access to the reference of correct fragment by using findFragmentByTag
method of the fragment manager.